+++ mobile/android/android-components/components/browser/tabstray/src/test/java/mozilla/components/browser/tabstray/TabTouchCallbackTest.kt
@@ -1,0 +1,82 @@
+/*
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ *  License, v. 2.0. If a copy of the MPL was not distributed with this
+ *  file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ */
+
+package mozilla.components.browser.tabstray
+
+import android.content.Context
+import android.support.v7.widget.helper.ItemTouchHelper
+import android.view.LayoutInflater
+import androidx.test.core.app.ApplicationProvider
+import mozilla.components.concept.tabstray.TabsTray
+import mozilla.components.support.base.observer.Observable
+import mozilla.components.support.base.observer.ObserverRegistry
+import mozilla.components.support.test.any
+import mozilla.components.support.test.mock
+import org.junit.Assert.assertEquals
+import org.junit.Assert.assertFalse
+import org.junit.Test
+import org.junit.runner.RunWith
+import org.mockito.Mockito.`when`
+import org.mockito.Mockito.never
+import org.mockito.Mockito.verify
+import org.robolectric.RobolectricTestRunner
+
+@RunWith(RobolectricTestRunner::class)
+class TabTouchCallbackTest {
+
+    private val context: Context
+        get() = ApplicationProvider.getApplicationContext()
+
+    @Test
+    fun `onSwiped notifies observers`() {
+        val observer: TabsTray.Observer = mock()
+        val observable: Observable<TabsTray.Observer> = ObserverRegistry()
+        observable.register(observer)
+
+        val touchCallback = TabTouchCallback(observable)
+        val viewHolder: TabViewHolder = mock()
+
+        touchCallback.onSwiped(viewHolder, 0)
+
+        verify(observer, never()).onTabClosed(any())
+
+        // With a session available.
+        `when`(viewHolder.session).thenReturn(mock())
+
+        touchCallback.onSwiped(viewHolder, 0)
+
+        verify(observer).onTabClosed(any())
+    }
+
+    @Test
+    fun `onChildDraw alters alpha of ViewHolder on swipe gesture`() {
+        val view = LayoutInflater.from(context).inflate(R.layout.mozac_browser_tabstray_item, null)
+        val holder = TabViewHolder(view, TabViewHolderTest.mockTabsTrayWithStyles())
+        val callback = TabTouchCallback(mock())
+
+        holder.itemView.alpha = 0f
+
+        callback.onChildDraw(mock(), mock(), holder, 0f, 0f, ItemTouchHelper.ACTION_STATE_DRAG, true)
+
+        assertEquals(0f, holder.itemView.alpha)
+
+        callback.onChildDraw(mock(), mock(), holder, 0f, 0f, ItemTouchHelper.ACTION_STATE_SWIPE, true)
+
+        assertEquals(1f, holder.itemView.alpha)
+    }
+
+    @Test
+    fun `alpha default is full`() {
+        val touchCallback = TabTouchCallback(mock())
+        assertEquals(1f, touchCallback.alphaForItemSwipe(0f, 0))
+    }
+
+    @Test
+    fun `onMove is not implemented`() {
+        val touchCallback = TabTouchCallback(mock())
+        assertFalse(touchCallback.onMove(mock(), mock(), mock()))
+    }
+}